Factor and add api_compare.py script
